@@ -13,9 +13,9 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
-        "oliverde8/php-etl-bundle": "^1.0",
-        "sylius/admin-bundle": "^1.08"
+        "php": ">=8.3",
+        "oliverde8/php-etl-bundle": "^v1.2.0-alpha1",
+        "sylius/admin-bundle": "^1.13"
     },
     "autoload": {
         "psr-4": {
