@@ -7,9 +7,7 @@ use Oliverde8\PhpEtlSyliusAdminBundle\Repository\Etl\EtlExecutionRepository;
 use Sylius\Component\Resource\Model\ResourceInterface;
 use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Entity(repositoryClass=EtlExecutionRepository::class)
- */
+#[ORM\Entity(repositoryClass:EtlExecutionRepository::class)]
 class EtlExecution extends BaseEtlExecution implements ResourceInterface
 {
     public function __construct()
