@@ -32,11 +32,11 @@ final class AdminMenuListener
 
         if ($executionsMenu = $menu->getChild('execution')) {
 
-            $dashboardMenu = $this->menuFactory->createItem('etl_executions_dashboard', ['route' => 'app_admin_etl_execution_dashboard'])
+            $dashboardMenu = $this->menuFactory->createItem('etl_executions_dashboard', ['route' => 'oliverde8_admin_etl_execution_dashboard'])
                 ->setLabel($this->translator->trans('sylius.ui.dashboard.title'))
                 ->setLabelAttribute('icon', 'options');
 
-            $gridMenu = $this->menuFactory->createItem('etl_executions_grid', ['route' => 'app_admin_etl_execution_index'])
+            $gridMenu = $this->menuFactory->createItem('etl_executions_grid', ['route' => 'oliverde8_admin_etl_execution_index'])
                 ->setLabel($this->translator->trans('sylius.ui.etl_executions'))
                 ->setLabelAttribute('icon', 'file');
 
