@@ -4,6 +4,7 @@ namespace Oliverde8\PhpEtlSyliusAdminBundle;
 
 use Oliverde8\PhpEtlSyliusAdminBundle\DependencyInjection\Oliverde8PhpEtlSyliusAdminExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class Oliverde8PhpEtlSyliusAdminBundle extends Bundle
@@ -16,7 +17,7 @@ class Oliverde8PhpEtlSyliusAdminBundle extends Bundle
         parent::build($container);
     }
 
-    public function getContainerExtension()
+    public function getContainerExtension(): ?ExtensionInterface
     {
         if (null === $this->extension) {
             $this->extension = new Oliverde8PhpEtlSyliusAdminExtension();
